bagian 10

+++ backend/routes/posts.js
@@ -38,7 +38,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/"+req.file.filename
+    imagePath: url + "/images/"+req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost =>{
     res.status(201).json({
@@ -48,7 +49,12 @@ router.post(
         id:createdPost._id,
       }
     });
-  });
+  })
+  .catch(error => {
+      res.status(500).json({
+        message: 'Creating a post failed!'
+      })
+  })
 });
 
 router.put("/:id", checAuth, multer({storage: storage}).single("image"),(req,res,next) =>{
@@ -61,11 +67,21 @@ router.put("/:id", checAuth, multer({storage: storage}).single("image"),(req,res
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
-  Post.updateOne({_id: req.params.id},post).then(result => {
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId},post).then(result => {
     console.log(result);
-    res.status(200).json({message: 'Update successful!'});
+    if(result.modifiedCount > 0){
+      res.status(200).json({message: 'Update successful!'});
+    }else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Couldn't update post!"
+    })
   });
 });
 
@@ -90,6 +106,11 @@ router.get('',(req,res,next) =>{
         posts: fetchedPosts,
         maxPosts: count
       });
+  })
+  .catch(error => {
+    res.status(500).json({
+        message: "Fetching post failed"
+    });
   });
 });
 
@@ -101,13 +122,26 @@ router.get('/:id',(req,res,next) =>{
        res.status(404).json({message : "Post not found!"})
     }
   })
+  .catch(error => {
+    res.status(500).json({
+        message: "Fetching post failed"
+    });
+  });
 })
 
 router.delete('/:id', checAuth, (req,res,next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post deleted!"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if(result.deletedCount > 0){
+      res.status(200).json({message: 'Deletion successful!'});
+    }else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   })
+  .catch(error => {
+    res.status(500).json({
+        message: "Fetching post failed"
+    });
+  });
 });
 
 module.exports = router;
